backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12
djangorestframework-simplejwt>=5.0
django-cors-headers>=3.14.0

psycopg2-binary>=2.9

celery>=5.2
redis>=4.0

requests>=2.26
drf-yasg>=1.21.4

pytest>=6.0
pytest-django>=4.5

pre-commit>=3.5.0
